MkMenuのitemを切り出して共通化
