+++ packages/frontend/src/style.scss
@@ -456,6 +456,10 @@ rt {
 	color: var(--MI_THEME-link);
 }
 
+._love {
+	color: var(--MI_THEME-love);
+}
+
 ._caption {
 	font-size: 0.8em;
 	opacity: 0.7;
